chore: refactor to ee module
