+++ app/services/action_service.rb
@@ -25,10 +25,6 @@ class ActionService
     @conversation.update!(priority: (priority[0] == 'nil' ? nil : priority[0]))
   end
 
-  def add_sla(sla_policy)
-    @conversation.update!(sla_policy_id: sla_policy.id)
-  end
-
   def add_label(labels)
     return if labels.empty?
 
@@ -93,3 +89,5 @@ class ActionService
     @conversation.additional_attributes['type'] == 'tweet'
   end
 end
+
+ActionService.include_mod_with('ActionService')
