+++ enterprise/app/services/action_service.rb
@@ -1,0 +1,5 @@
+class ActionService
+  def add_sla(sla_policy)
+    @conversation.update!(sla_policy_id: sla_policy.id)
+  end
+end
