+++ spec/services/action_service_spec.rb
@@ -45,15 +45,4 @@ describe ActionService do
       expect(conversation.reload.team).to be_nil
     end
   end
-
-  describe '#add_sla' do
-    let(:sla_policy) { create(:sla_policy, account: account) }
-    let(:conversation) { create(:conversation, account: account) }
-    let(:action_service) { described_class.new(conversation) }
-
-    it 'adds the sla policy to the conversation' do
-      action_service.add_sla(sla_policy)
-      expect(conversation.reload.sla_policy).to eq(sla_policy)
-    end
-  end
 end
